Established keys  to puzzles in game.
Creted/modified the two doors that will act as the puzzles in the game. Still working on making keypad (combination lock) work after reading the Inform 7 handbook. Added object that will give hint to the door code. Rearranged rooms to start in economy, game will focus more on getting from economy to first class + galley and eventually in to cockpit, less focus on "killing" (or restraining?) terrorist to make game suitable for a wider audience.